Moved some code to new ApiInterface file
